+++ src/error_chain.rs
@@ -166,11 +166,17 @@ macro_rules! impl_error_chain_processed {
                 where F: FnOnce() -> EK, EK: Into<$error_kind_name> {
                 $error_name::with_chain(self, Self::from_kind(error().into()))
             }
+
+            /// A short description of the error.
+            /// This method is identical to [`Error::description()`](https://doc.rust-lang.org/nightly/std/error/trait.Error.html#tymethod.description)
+            pub fn description(&self) -> &str {
+                self.0.description()
+            }
         }
 
         impl ::std::error::Error for $error_name {
             fn description(&self) -> &str {
-                self.0.description()
+                self.description()
             }
 
             #[allow(unknown_lints, unused_doc_comment)]
@@ -236,14 +242,6 @@ macro_rules! impl_error_chain_processed {
         impl From<String> for $error_name {
             fn from(s: String) -> Self {
                 $error_name::from_kind(s.into())
-            }
-        }
-
-        impl ::std::ops::Deref for $error_name {
-            type Target = $error_kind_name;
-
-            fn deref(&self) -> &Self::Target {
-                &self.0
             }
         }