+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - [Remove `impl Deref<Kind> for Error`](https://github.com/rust-lang-nursery/error-chain/pull/192)
+- [Shrink the size of `Error` to a pointer](https://github.com/rust-lang-nursery/error-chain/pull/225)
 
 # 0.11.0